requirements.txt
fastapi
uvicorn
streamlit
faiss-cpu
numpy
pandas
requests
bs4
python-dotenv
google-generativeai
